+++ app/src/main/java/com/techcourse/support/web/mapping/HandlerMappings.java
@@ -1,6 +1,5 @@
 package com.techcourse.support.web.mapping;
 
-import com.techcourse.ManualHandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
